+++ bocadillo/websockets.py
@@ -1,11 +1,16 @@
-from typing import Any, Awaitable, Callable, Optional, Tuple, Union
+import traceback
+from typing import Any, Callable, Tuple, Union
 
+import typesystem
 from starlette.datastructures import URL, Headers, QueryParams
 from starlette.websockets import WebSocket as StarletteWebSocket
 from starlette.websockets import WebSocketDisconnect as _WebSocketDisconnect
 
 from .app_types import Event, Receive, Scope, Send
+from .compat import asyncnullcontext
 from .constants import WEBSOCKET_CLOSE_CODES
+from .converters import ViewConverter, convert_arguments
+from .injection import consumer
 
 
 class WebSocket:
@@ -224,5 +229,34 @@ class WebSocket:
             yield await self.receive()
 
 
-WebSocketView = Callable[[WebSocket], Awaitable[None]]
+class WebSocketConverter(ViewConverter):
+    def get_query_params(self, args: tuple, kwargs: dict) -> dict:
+        ws: WebSocket = args[0]
+        return ws.query_params
+
+
+class WebSocketView:
+
+    __slots__ = ("func",)
+
+    def __init__(self, func: Callable):
+        func = convert_arguments(func, converter_class=WebSocketConverter)
+        func = consumer(func)
+        self.func = func
+
+    async def __call__(self, ws: WebSocket, **params):
+        context = ws if ws.auto_accept else asyncnullcontext()
+        try:
+            async with context:
+                try:
+                    await self.func(ws, **params)  # type: ignore
+                except typesystem.ValidationError:
+                    await ws.ensure_closed(403)
+                    traceback.print_exc()
+        except BaseException:
+            traceback.print_exc()
+            await ws.ensure_closed(1011)
+            raise
+
+
 WebSocketDisconnect = _WebSocketDisconnect